Former-commit-id: 90bb70f372b0ca3ac466944da797ce849842fb7a
Former-commit-id: 5b2d80bf5ebcbaedb242f5b0119b0b6f9c2975bd

+++ server/src/main/java/org/icgc/dcc/validation/cascading/RemoveHeaderFilter.java
@@ -29,7 +29,7 @@ public class RemoveHeaderFilter extends BaseOperation implements Filter {
 
   @Override
   public boolean isRemove(FlowProcess flowProcess, FilterCall filterCall) {
-    if(filterCall.getArguments().getInteger(ValidationFields.OFFSET_FIELD_NAME) == 0) {
+    if(filterCall.getArguments().getLong(ValidationFields.OFFSET_FIELD_NAME) == 0) {
       return true;
     } else {
       return false;
